services/system/requirements.txt
paho-mqtt
psutil
python-dotenv
